fix: unit test process not exiting at the end
